+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.80 (Pending):
 * Update AbstractServerTest to support binding EntityManager under alternative name.
+* Update AbstractServerTest to add helper methods that iterate over all EntityManager
+  and begin, commit and rollback transactions on EntityManager instances.
 
 ## 0.79:
 * Update PersistenceTestModule to support binding under alternative name.